refactor(thunk): change the actionCreator to make the async call and return a thunk
a thunk is a function that returns a function. if the thunk also has a 'dispatch' param the inner
reurned function can all 'dispatch'. Its beautiful! redux-thunk 'sees' the actionCreator returns a
function and steps in to cll it with dispatch. Also, make the dispatch call in the component
explicit.